Display mode MUST not be specified in the print method
As it could overwite user specified display mode.

Therefore, since it needs to be specifed so that any text could be
displayed on the screen, and it is not desired to force user to specify
it manually every time, it is better to set it to normal mode by default
in the reset method, when device is being reseted.

It is also possible to overwite display mode by user.
Also that user wouldn't have to specify it if all they want is normal
mode by default which is set by the library for the user.

Now, let's get ready for the next change and next commit as well :D